requirements.txt
requests==2.20.1
python-daemon==2.2.3
